refactor(aegisctl): remove legacy inject submit/metadata/describe/logs
The backend accepts guided configs exclusively (see AegisLab PR #74), so the
`inject submit`, `inject metadata`, `inject describe`, and `inject logs`
commands no longer map to live endpoints. Drop them along with their
supporting types (`InjectSpec`, `FaultSpec`, `fieldDescriptionCLI`) and
the unused `runInjectSubmitWait` / `runInjectWait` machinery. Move the
still-used types (`injectSubmitResponse`, `injectSubmitItem`, `ContainerRef`,
`LabelItem`, `ParameterSpec`) into a shared `inject_types.go` so
`inject guided --apply`, `regression run`, and the execute CLI keep
compiling. Update `root` and `wait` help text to point at `inject guided`.
